refactor jupyter code to jupyter_gui.py and update API
